Add Collision Tree-Shuffling in SimpleGame (read log desc. plz)
My git logs have been pretty terrible, so I don't actually know what the
state of the project the last time I made a commit; The most recent log I
made was changing the Collision OctTree so that it reshuffles actors that
have moved, rather than just rebuilding the tree every frame. However, I
do not know if the last time I commited, the thing even had OctTrees.
Henceforce, I'll try to be better with logs (even though I'm the only one
reading these) but for now, the current state of this Repo is thus:

GL_Invaders hasn't been changed in forever, it's basically all I needed it
to be

GL_SkeletalAnimationTest currently is able to read SkeletalMeshes from
Collada .dae files (if they are formatted as Blender's output script
formatts them), and play a stored animation with SLERP between frames.
(uses quaternians for rotation).

SimpleGame has a large number of randomly placed red boxes, one fifth of
which move, and one of which is controlled with WASD and the mouse. They
have their collision checked against each other using an OctTree (also,
pressing 'b' will switch to a standard O(n^2) collision check of all the
boxes, for comparison). The boxes turn blue if moving, yellow if in
contact with another box, and green for 1 frame when switching their node
in the OctTree.

My plan is to keep adding functionallity to SimpleGame, as well as adding
bits to GL_SkeletalAnimationTest when I'm tired of SimpleGame. We'll see
what happens.